// src/app/modules/push/model/device.model.ts
export class Device {
    devName: string="";
    devMode: string="";
    devStatus: string="";
    dvVlan: string="";
    devDesc: any;
    endptsVlan: any;
    cdpPlatform: any;
    lldpSysDesc: any;
    codingStatus: string="";
    servTemplate: string="";
  }
  

  
// interface Device {
//   devName: string;
//   devMode: string;
//   devStatus: string;
//   dvVlan: string;
//   devDesc: any;
//   endptsVlan: any;
//   cdpPlatform: any;
//   lldpSysDesc: any;
//   codingStatus: string;
//   servTemplate: string;
// }